requirements.txt
bcrypt==4.0.1
cryptography==41.0.1
pycryptodome==3.19.0
argon2-cffi==21.3.0
pyperclip==1.8.2 